Fixed issue where could not derive shard form worker name.
This was caused by a over restrictive regular exporession used to
extract the shard, which depended on the name of service. When the name
of service changed due to helm namespacing, it could not extract the
shard. Made the regular exression less dependent on service name.
Now it only requires workers to be named worker-n where is the worker
shard.