Configure deployments and services to use selectors macro.
This allows the selectors to be configured using the macro, seperate
from the labels macro. The selectors in then macro are also less
restrictive, for example it does not check for app version or
deployment envrionment. This is important to ensure that helm upgrade
across app versions/deployment envrionment does not break the app.